# Conflicts:
#	libnd4j/include/ops/declarable/generic/recurrent/lstmBlockCell.cpp
#	libnd4j/include/ops/declarable/headers/recurrent.h
#	libnd4j/include/ops/declarable/helpers/cpu/lstm.cpp

+++ nd4j/nd4j-backends/nd4j-tests/src/test/java/org/nd4j/linalg/rng/RandomTests.java
@@ -214,7 +214,7 @@ public class RandomTests extends BaseNd4jTest {
     public void testLinspace1() {
         INDArray z1 = Nd4j.linspace(1, 100, 200, DataType.DOUBLE);
 
-        Linspace linspace = new Linspace(1, 100, 200, DataType.DOUBLE);
+        Linspace linspace = new Linspace((double) 1, (double) 100, 200, DataType.DOUBLE);
         Nd4j.getExecutioner().exec(linspace, Nd4j.getRandom());
 
         INDArray z2 = linspace.z();
